// mat-project-vite-test/src/api/dtos/task/take/response.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Exercise = {
  type?: "exercise";
  exerType: string;
  instructions: {
    content: string;
    [k: string]: unknown;
  };
  content: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
} & (TakeResponse | TakeResponse1);

export interface TakeTaskResponse {
  data: {
    task: {
      name: string;
      display?: "vertical" | "horizontal";
      description: string;
      entries: (Group | Exercise)[];
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface Group {
  type?: "group";
  resources: {
    content: string;
    [k: string]: unknown;
  }[];
  entries: (Exercise | Group)[];
  [k: string]: unknown;
}
export interface TakeResponse {
  exerType: "Doplnovacka";
  content: (Cmb | TxtI | string)[];
  [k: string]: unknown;
}
export interface Cmb {
  /**
   * @minItems 1
   */
  values: [
    {
      [k: string]: unknown;
    },
    ...{
      [k: string]: unknown;
    }[]
  ];
  selectedIndex?: number | null;
  [k: string]: unknown;
}
export interface TxtI {
  value?: string | null;
  [k: string]: unknown;
}
export interface TakeResponse1 {
  exerType: "HledaniChyb";
  content: {
    defaultText: string;
    text?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}